backend/app/utils/image_analysis.py
import os
import random
import numpy as np
import io
from PIL import Image

# In a real implementation, we would use a proper facial emotion detection model
# such as DeepFace or a pre-trained model from HuggingFace

# For demonstration purposes, we'll create a simple mock implementation
# In production, replace this with actual model inference

def analyze_face_emotion(image_bytes):
    """
    Analyze the emotion from a facial image.
    
    Args:
        image_bytes (bytes): The image file bytes
        
    Returns:
        tuple: (emotion, score) where emotion is a string and score is a float 0-100
    """
    # In a real implementation, we would use:
    # 1. DeepFace
    # from deepface import DeepFace
    # img = Image.open(io.BytesIO(image_bytes))
    # img.save('temp.jpg')
    # result = DeepFace.analyze('temp.jpg', actions=['emotion'])
    # dominant_emotion = result[0]['dominant_emotion']
    # 
    # 2. Or a pre-trained model from HuggingFace
    # from transformers import pipeline
    # emotion_classifier = pipeline('image-classification', model='joeddav/vit-base-patch16-224-in21k-finetuned-emotion')
    # img = Image.open(io.BytesIO(image_bytes))
    # result = emotion_classifier(img)
    # dominant_emotion = result[0]['label']
    
    # For demo purposes, we'll randomly select an emotion
    emotions = ["happy", "sad", "angry", "neutral", "surprised", "fearful"]
    weights = [0.3, 0.2, 0.1, 0.2, 0.1, 0.1]  # Weighted probabilities
    
    # Try to open the image to verify it's valid
    try:
        img = Image.open(io.BytesIO(image_bytes))
        # In a real implementation, we would preprocess the image here
        # img = img.resize((224, 224))  # Resize for model input
        # img_array = np.array(img)  # Convert to numpy array for model
    except Exception as e:
        raise ValueError(f"Invalid image: {str(e)}")
    
    # Randomly select an emotion based on weights
    dominant_emotion = random.choices(emotions, weights=weights, k=1)[0]
    
    # Generate a score (0-100) with some randomness but weighted toward the emotion
    base_score = random.randint(70, 95)  # Base score between 70-95
    variation = random.randint(-10, 10)  # Add some variation
    score = min(max(base_score + variation, 0), 100)  # Ensure score is between 0-100
    
    return dominant_emotion, score